remove barrel file
